Don't install git and xet on R3.6.3 image
This error message appears:

git : Depends: libc6 (>= 2.38) but 2.28-10+deb10u2 is to be installed Depends: libpcre2-8-0 (>= 10.34) but 10.32-5+deb10u1 is to be installed

and apparently

Please note that upgrading libc6 can be risky, as it's a core library that many other packages depend on. Be careful to avoid breaking your system.

So, if we really need xet, then we can do this using a different image.